+++ codes/project/student.h
@@ -13,7 +13,7 @@ typedef struct {
 void    setNum(size_t);
 size_t  getNum(void);
 void    setNamelen(int);
-size_t  getNamelen(void);
+int     getNamelen(void);
 void    setStudentPtr(Student *);
 Student *getStudentPtr(void);
